requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
feedparser==6.0.11
newspaper3k==0.2.8
pytz==2024.1
lxml==5.3.0
beautifulsoup4==4.12.3
requests==2.32.3
